requirements.txt
typer
requests
beautifulsoup4
socket
json
python-dotenv
python-whois
shodan
dnspython